pyproject.toml
[project]
name = "fedga"
version = "0.1.0"
description = "Federated Leanring with Gradient Alignment"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "datasets==2.21.0",
    "python-box==7.0.0",
    "scipy==1.12.0",
    "torchvision==0.17.0",
    "torchmetrics==0.10.1",
    "wandb==0.17.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint.isort]
known-third-party = ["wandb"]
